move most of main to a new, different package
It turns out you can't import `main` from another repo.
